Reload Eel on Python file changes
Add support for reloading the Bottle server during development. Eel on
the JS side will now try to automatically reconnect to the Python/Bottle
server when the websocket dies. This allows us to let the Bottle server
die and restart to pull in new changes.

An explicit port must be set when we want to use the reloading server to
make sure that it restarts on the same port, as that is the port that
the JS side will be trying to connect to.